Results so far
The Magnus Tools Jupyter notebook describes results that have been obtained so far from the Magnus Expansion project.  Note that some parts of the notebook may not run because they contain example code that is prohibited in Jupyter (for example, defining functions in MATLAB is forbidden but the notebook includes the code for some functions).  The notebook also requires some supporting code that I have yet to upload.

The Spin Simulations (presentation mode) describes my results from Summer 2020 considering the fidelity of various pulse sequences over a 4-spin system.  This notebook is configured to be run as a presentation using the RISE extension to allow code to be run while presenting (to show plots in 3D, for example).