Pulse-length correction to AHT
The spacing between instantaneous frame changes in the AHT calculation has been changed.  Half the pulse length (1.4 microseconds by default) is allocated to the spacing on either side of the pulse (so the spacing between two pulses is now tau+p1, with p1/2 added instead to the beginning and end).