New Results; utility functions
Magnus Tools now has new results using the proper pulse-length correction introduced in the latest version of fullMagnusOneParam (scroll all the way to the bottom of the notebook to find these results).

Uploaded a number of utility functions that make some of the notebooks (particularly One Parameter SpinSim) work.  Most of these are also defined as built-in functions in fullMagnusOneParam.