Organization Update; additional metrics
There are two parts to this update.  First, all code files now specify the author and date of creation in a uniform format.  Second, the propagator distance (Hilbert-Schmidt distance) has been added as a secondary metric.  The metric is computed in uDist.m and is saved in runAnalysis alongside the overlap metric, which has been renamed from metric.m to overlap.m.